+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- added: Support for hyperevm to ChangeHero, Exolix, LetsExchange.
+
 ## 2.28.0 (2025-07-17)
 
 - changed: Remove Thorchain minimum amounts from the thrown `SwapBelowLimit` errors